src/backend/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-openai
pydantic
python-dotenv
